Fix fnmut problem
Apparently, the issue was that the function itself was viewed as being
"mutated," which makes sense if you think about the state of the random
number generator as being part of the function itself. Pretty neat, to
be honest.

Anyway, what this means is that the function needs to be declared as
mutable in the declaration for the consuming function.